+++ ros/.gitignore
@@ -1,3 +1,4 @@
 .catkin_workspace
 *.swp
 **/__pycache__
+**/*.pyc
