--- ros/src/i2c_com/README.txt
@@ -1,9 +1,0 @@
-Welcome to i2c_comm, the i2c communication package.
-
-This package handles communication to the i2c hardware. This package
-does not do any processing and merely relays data onto ROS.
-
-Nodes:
-  * i2c_comm:
-    -subs: i2c_tx
-    -pubs: i2c_rx
