--- ros/src/i2c_com/package.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-    <name>i2c_com</name>
-    <version>0.0.0</version>
-    <description>The i2c_com package</description>
-    <maintainer email="todoh@todo.todo">TODO</maintainer>
-
-    <license>TODO</license>
-
-    <buildtool_depend>catkin</buildtool_depend>
-    <build_depend>shared_msgs</build_depend>
-    <exec_depend>shared_msgs</exec_depend>
-
-</package>
