+++ ros/src/utils/package.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0"?>
+<package format="2">
+    <name>utils</name>
+    <version>0.0.0</version>
+    <description>The utils package</description>
+
+    <!-- One maintainer tag required, multiple allowed, one person per tag -->
+    <!-- Example:  -->
+    <!-- <maintainer email="jane.doe@example.com">Jane Doe</maintainer> -->
+    <maintainer email="ivan@todo.todo">keith</maintainer>
+
+
+    <!-- One license tag required, multiple allowed, one license per tag -->
+    <!-- Commonly used license strings: -->
+    <!--   BSD, MIT, Boost Software License, GPLv2, GPLv3, LGPLv2.1, LGPLv3 -->
+    <license>TODO</license>
+
+    <buildtool_depend>catkin</buildtool_depend>
+    <build_depend>rospy</build_depend>
+    <build_depend>shared_msgs</build_depend>
+    <build_depend>std_msgs</build_depend>
+    <build_export_depend>rospy</build_export_depend>
+    <build_export_depend>sensor_msgs</build_export_depend>
+    <build_export_depend>shared_msgs</build_export_depend>
+    <build_export_depend>std_msgs</build_export_depend>
+    <exec_depend>rospy</exec_depend>
+    <exec_depend>shared_msgs</exec_depend>
+    <exec_depend>std_msgs</exec_depend>
+
+
+    <!-- The export tag contains other, unspecified, tags -->
+    <export>
+        <!-- Other tools can request additional information be placed here -->
+
+    </export>
+</package>
