+++ surface/compros.sh
@@ -1,0 +1,6 @@
+#! /bin/bash
+
+export ROS_IP=192.168.1.11
+export ROS_MASTER_URI=http://192.168.1.3:11311
+export ROS_HOSTNAME=192.168.1.11
+
